@@ -24,7 +24,7 @@
   "author": "https://github.com/anorudes, https://github.com/keske",
   "license": "MIT",
   "devDependencies": {
-    "autoprefixer-core": "^5.2.1",
+    "autoprefixer-core": "^6.0.1",
     "babel-core": "^5.4.7",
     "babel-eslint": "^4.0.7",
     "babel-loader": "^5.1.4",
@@ -32,11 +32,11 @@
     "bootstrap": "^3.3.5",
     "bootstrap-webpack": "0.0.3",
     "classnames": "^2.1.3",
-    "css-loader": "^0.15.6",
+    "css-loader": "^0.18.0",
     "cssnext-loader": "^1.0.1",
-    "csswring": "^3.0.5",
+    "csswring": "^4.0.0",
     "eslint": "^1.1.0",
-    "eslint-config-airbnb": "0.0.7",
+    "eslint-config-airbnb": "0.0.8",
     "eslint-plugin-react": "^3.2.3",
     "extract-text-webpack-plugin": "^0.8.2",
     "file-loader": "^0.8.4",
@@ -47,21 +47,26 @@
     "jquery": "^2.1.4",
     "jsx-control-statements": "^1.1.1",
     "node-sass": "^3.1.2",
-    "postcss-loader": "^0.5.1",
-    "react": "^0.13.3",
+    "postcss-loader": "^0.6.0",
+    "react": "^0.14.0-rc1",
+    "react-dom": "^0.1.0",
     "react-hot-loader": "^1.2.7",
-    "react-redux": "^0.2.2",
-    "react-router": "1.0.0-beta3",
-    "redux": "^1.0.1",
-    "redux-devtools": "^0.1.3",
-    "redux-logger": "0.0.3",
-    "redux-react-router": "^0.2.1",
+    "react-redux": "^2.1.2",
+    "react-router": "1.0.0-rc1",
+    "redux": "^3.0.0",
+    "redux-devtools": "^2.1.2",
+    "redux-logger": "1.0.6",
+    "redux-react-router": "^1.0.0-beta1",
     "redux-thunk": "^0.1.0",
-    "sass-loader": "^1.0.2",
+    "sass-loader": "^2.0.1",
     "style-loader": "^0.12.3",
     "url-loader": "^0.5.6",
     "webpack": "^1.9.6",
     "webpack-dev-server": "^1.8.2",
     "yargs": "^3.18.0"
+  },
+  "dependencies": {
+    "react": "^0.14.0-rc1",
+    "react-dom": "^0.14.0-rc1"
   }
 }
