@@ -1,4 +1,5 @@
 import React from 'react';
+import ReactDOM from 'react-dom';
 import App from './App.js';
 
-React.render(<App />, document.getElementById('App'));
+ReactDOM.render(<App />, document.getElementById('App'));
